+++ lib/dicey/sum_frequency_calculators/multinomial_coefficients.rb
@@ -72,7 +72,7 @@ module Dicey
       # @return [Array<Integer>]
       def next_row_of_coefficients(row_index, window_size, previous_row)
         length = (row_index * window_size) + 1
-        (0..length).map do |col_index|
+        (0...length).map do |col_index|
           # Have to clamp to 0 to prevent accessing array from the end.
           # BUG: TruffleRuby can't handle endless range in #clamp (see https://github.com/oracle/truffleruby/issues/3945)
           window_range = ((col_index - window_size).clamp(0..col_index)..col_index)