+++ lib/dicey/sum_frequency_calculators/brute_force.rb
@@ -28,44 +28,10 @@ module Dicey
 
       def calculate(dice, **nil)
         dice = vectorize_dice(dice) if defined?(VectorNumber)
-        combine_dice_enumerators(dice.map(&:sides_list)).map(&:sum).tally
+        dice.map(&:sides_list).reduce { |result, die|
+          result.flat_map { |roll| die.map { |side| roll + side } }
+        }.tally
       end
-
-      if defined?(Enumerator::Product)
-        # Get an enumerator which goes through all possible permutations of dice sides.
-        #
-        # @param side_lists [Enumerable<Enumerable<Any>>]
-        # @return [Enumerator<Enumerable<Enumerable<Any>>>]
-        def combine_dice_enumerators(side_lists)
-          Enumerator::Product.new(*side_lists)
-        end
-      # :nocov:
-      else
-        # Get an enumerator which goes through all possible permutations of dice sides.
-        #
-        # @param side_lists [Enumerable<Enumerable<Any>>]
-        # @return [Enumerator<Array<Array<Any>>>]
-        def combine_dice_enumerators(side_lists)
-          product(side_lists)
-        end
-
-        # Simplified implementation of {Enumerator::Product}.
-        # Adapted from {https://bugs.ruby-lang.org/issues/18685#note-10}.
-        #
-        # @param enums [Enumerable<Enumerable<Any>>]
-        # @return [Enumerator<Array<Array<Any>>>]
-        def product(enums, &block)
-          return to_enum(__method__, enums) unless block_given?
-
-          enums
-            .reverse
-            .reduce(block) { |inner, enum|
-              ->(values) { enum.each_entry { inner.call([*values, _1]) } }
-            }
-            .call([])
-        end
-      end
-      # :nocov:
     end
   end
 end