+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Next]
 
+**Changed**
+- `KroneckerSubstitution` calculator now accepts dice with all integers (including negative). Additionally, performance now depends on the total range of numbers instead of the maximum number.
+
 [Compare v0.16.0...main](https://github.com/trinistr/dicey/compare/v0.16.0...main)
 
 ## [v0.16.0] — 2025-10-09